survey_simulation/requirements.txt
openai-agents
python-dotenv
matplotlib
numpy
pandas
seaborn
scipy 